Overcloud host and service configuration changes required to add a set of Cascade Lake compute nodes in addition to the existing Skylake compute nodes
The CCLake nodes will be running in dedicated aggregates with hyperthreading enabled; one aggregate for Gaia projects where pinned CPUs are not used and a larger aggregate for EUCLID where NUMA awareness and CPU pinning are used.

To accomplish this, nova.conf overrides are fed through for the hosts in the different aggregates. There may be a smarter way to achieve this than a a host-specific override.

Additionally, role playbooks are included for enabling huge pages, OFED and SRIOV on the CCLake hosts.